+++ dev-packages/browser-integration-tests/suites/tracing/user-timing/disabled/init.js
