+++ dev-packages/browser-integration-tests/suites/tracing/user-timing/pageload-measure-spans-domexception-details/init.js
