+++ dev-packages/browser-integration-tests/suites/tracing/user-timing/pageload-measure-spans/init.js
