+++ dev-packages/browser-integration-tests/suites/tracing/user-timing/spans/init.js
